Module 4 homework -changed GetCommission() return type to double in the Bank class  -rewrote methods in BankSystemImpl class more efficiently  -rewrote methods in ChinaBank, EUBank, USBank more efficiently  -enum Currency values are now one per line  -used UUID.randomUUID() to autogenerate positive Id's in the Main class  -overriden toString(), equals(), hashCode() methods in the User class

+++ src/module4/homework/Bank.java
@@ -27,13 +27,11 @@ public abstract class Bank {
 
     public abstract int getMonthlyRate();
 
-    public abstract int getCommission(int sum);
+    public abstract double getCommission(int sum);
 
     public double moneyPaidMonthlyForSalary() {
         return 1000.0;
     }
-
-    ;
 
     public long getId() {
         return id;
